refactor: move all the loss functions from  common_metrics.py file into loss.py and  move training files from models/ to data/ and update imports
