+++ training/model_training.py
@@ -12,14 +12,14 @@ import numpy as np
 import albumentations as A
 from tqdm import tqdm
 import gc
-from common_metrics import validate_all, record_validation_metrics_to_csv, getLossFunction
-from models_tcloud import save_model_and_log_params, dict_to_hash_key,init_model_and_loaders
+from models.common_metrics import validate_all, record_validation_metrics_to_csv, getLossFunction
+from models.models_tcloud import save_model_and_log_params, dict_to_hash_key,init_model_and_loaders
 import torch.nn.functional as F
 import pandas as pd
 import segmentation_models_pytorch as smp
 from libraries.utils import save_geotiff, write_dict_to_json
 import json
-from model_test import evaluate_on_test_set
+from models.model_test import evaluate_on_test_set
 from libraries.wandb_retrieve import wandinit
 from libraries.utils import get_preds_multi_encoders, set_seed
 import random
